+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added `table.types.CastTo(from, to)` helper for cast `types.Value` to golang native type value destination
+
 ## v3.38.3
 * Added retries to initial discovering